requirements.txt
discord.py==2.1.0
langchain==0.0.130
openai==0.27.2
openapi-schema-pydantic==1.2.4
pip-chill==1.0.1
pytelegrambotapi==4.10.0
wincertstore==0.2
